modularise macos native code
